Infrastructure: refactor the custom error pages module
This commit refactors the custom error pages module to make it more
compliant with standard golang practices, optimize performance and avoid
the need for externally served resources (the returned page embeds all
necessary resources).
