+++ infrastructure/ingress-controller/custom-error-pages/README.md
@@ -3,32 +3,24 @@
 This folder contains the material necessary for the creation of the service used by the `nginx ingress controller` to serve custom error pages.
 
 ## Main components
+
 The core components required to serve custom error pages are:
 
-* [main.go](main.go): the source code of the server in charge of returning the desired error page, together with the companion static resources. The error page is automatically customized depending on the headers configured by the ingress controller during the request (i.e. the error code and the format requested);
-* [rootfs](rootfs): the resources to be embedded into the docker image and used by the server. In detail:
-
-    * [www/templates](rootfs/www/templates) contains the template pages which are filled in by the server with the error information before being returned;
-    * [www/static](rootfs/www/static) contains the static companion resources to be served (e.g. css and images).
+* [main.go](server/main.go): the source code of the server in charge of returning the desired error page. The error page is automatically customized depending on the headers configured by the ingress controller during the request (i.e. the error code and the format requested);
+* [templates](static/templates): the template pages which are filled in by the server with the error information before being returned. The original, non-minified version of the templates is available in the [original](static/original) directory.
 
 ## Customization
-The behavior of the server can be customized by setting different environment variables:
 
-* `ERROR_SERVER_PORT`: the port the `http` server is listening to (default: `8080`);
-* `ERROR_TEMPLATES_PATH`: the directory within the Docker image containing the error templates (default: `/www/templates`);
-* `ERROR_STATIC_PATH`: the directory within the Docker image containing the static companion resources  (default: `/www/static`);
-* `ERROR_STATIC_SERVE_PATH`: the path where the static resources are served by the handler (default: `/error-page`);
-* `ERROR_STATIC_URI`: the relative or absolute pointing to the location where the static resources are made accessible (default: `/error-page`). Use an absolute URL in case multiple domains are served by the ingress controller.
+The behavior of the server can be customized through the following command line parameters:
+
+* `--http-address`: the address the server binds to (defaults to `:8080`);
+* `--templates-path`: the directory within the Docker image containing the error templates (defaults to `/templates`);
 
 ## How to build
-The creation of the Docker image is automatized through the [Makefile](Makefile). Please, customize the repository configuration according to your requirements.
 
-**TL;DR:**
 ```bash
-$ make
-$ make push
+docker build . --push -tag crownlabs/custom-error-pages:v0.x
 ```
-
 
 ## Additional information
