+++ infrastructure/ingress-controller/custom-error-pages/go.mod
@@ -1,0 +1,8 @@
+module github.com/netgroup-polito/CrownLabs/infrastructure/ingress-controller/custom-error-pages
+
+go 1.16
+
+require (
+	github.com/prometheus/client_golang v1.11.0
+	k8s.io/klog/v2 v2.9.0
+)
