--- infrastructure/ingress-controller/custom-error-pages/rootfs/www/static/error-page.css
@@ -1,128 +1,0 @@
-* {
-  -webkit-box-sizing: border-box;
-  box-sizing: border-box;
-}
-
-body {
-  padding: 0;
-  margin: 0;
-}
-
-#notfound {
-  position: relative;
-  height: 100vh;
-}
-
-#notfound .notfound {
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  -webkit-transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%);
-  transform: translate(-50%, -50%);
-}
-
-.notfound {
-  max-width: 520px;
-  width: 100%;
-  line-height: 1.4;
-  text-align: center;
-}
-
-.notfound .notfound-err {
-  position: relative;
-  height: 200px;
-  margin: 0px auto 20px;
-  z-index: -1;
-}
-
-.notfound .notfound-err h1 {
-  font-family: 'Montserrat', sans-serif;
-  font-size: 236px;
-  font-weight: 200;
-  margin: 0px;
-  color: #211b19;
-  text-transform: uppercase;
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  -webkit-transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%);
-  transform: translate(-50%, -50%);
-}
-
-.notfound .notfound-err h2 {
-  font-family: 'Montserrat', sans-serif;
-  font-size: 28px;
-  font-weight: 400;
-  text-transform: uppercase;
-  color: #211b19;
-  background: #fff;
-  padding: 10px 5px;
-  margin: auto;
-  display: inline-block;
-  position: absolute;
-  bottom: 0px;
-  left: 0;
-  right: 0;
-}
-
-.notfound img {
-  width: 400px;
-  margin-bottom: 25px;
-}
-
-.notfound a {
-  font-family: 'Montserrat', sans-serif;
-  display: inline-block;
-  font-weight: 700;
-  text-decoration: none;
-  color: #211b19;
-  text-transform: uppercase;
-  margin-top: 25px;
-  padding: 13px 23px;
-  border: 2px solid #211b19;
-  border-radius: 25px;
-  font-size: 24px;
-  -webkit-transition: 0.2s all;
-  transition: 0.2s all;
-}
-
-.notfound a:hover {
-  color: #fff;
-  background: #211b19;
-}
-
-@media only screen and (max-width: 767px) {
-  .notfound .notfound-err h1 {
-    font-size: 148px;
-  }
-  .notfound img {
-    width: 300px;
-    margin-bottom: 15px;
-  }
-  .notfound a {
-    margin-top: 15px;
-  }
-}
-
-@media only screen and (max-width: 480px) {
-  .notfound .notfound-err {
-    height: 148px;
-    margin: 0px auto 10px;
-  }
-  .notfound .notfound-err h1 {
-    font-size: 86px;
-  }
-  .notfound .notfound-err h2 {
-    font-size: 16px;
-  }
-  .notfound img {
-    width: 200px;
-    margin-bottom: 0px;
-  }
-  .notfound a {
-    padding: 7px 15px;
-    font-size: 14px;
-  }
-}
