--- infrastructure/ingress-controller/custom-error-pages/rootfs/www/templates/error-page.html.tmpl
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="utf-8">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-
-	<title>CrownLabs</title>
-
-	<!-- Google font -->
-	<link href="https://fonts.googleapis.com/css?family=Montserrat:200,400,700" rel="stylesheet">
-
-	<!-- Custom stlylesheet -->
-	<link type="text/css" rel="stylesheet" href="{{.StaticUri}}/error-page.min.css" />
-
-	<!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
-	<!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-	<!--[if lt IE 9]>
-		  <script src="https://oss.maxcdn.com/html5shiv/3.7.3/html5shiv.min.js"></script>
-		  <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
-		<![endif]-->
-
-</head>
-
-<body>
-	<div id="notfound">
-		<div class="notfound">
-			<img src="{{.StaticUri}}/crown.svg" />
-			<div class="notfound-err">
-				<h1>Oops!</h1>
-				<h2>{{.ErrorCode}} &ndash; {{.ErrorMsg}}</h2>
-			</div>
-			<a href="/">Homepage</a>
-		</div>
-	</div>
-</body><!-- This templates was made by Colorlib (https://colorlib.com) -->
-
-</html>
