+++ infrastructure/ingress-controller/custom-error-pages/static/original/error-page.json.tmpl
