+++ infrastructure/ingress-controller/custom-error-pages/static/templates/error-page.json.tmpl
@@ -1,0 +1,1 @@
+{"error":{"code":{{.ErrorCode}},"message":"{{.ErrorMsg}}"}}
